Create flag for stdtest, Make flake8 happy
